+++ elmyra.ip.access.epo/elmyra/ip/access/epo/services/util.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# (c) 2013-2015 Andreas Motl, Elmyra UG
+# (c) 2013-2016 Andreas Motl, Elmyra UG
 import re
 import json
 import logging
@@ -159,3 +159,18 @@ def numberlist_util_handler(request):
 
     return response
 
+def request_to_options(request, options):
+
+    # TODO: transfer all modifiers 1:1
+
+    if asbool(request.params.get('query_data[modifiers][family-remove]')):
+        options.update({'feature_family_remove': True})
+    elif asbool(request.params.get('query_data[modifiers][family-replace]')):
+        options.update({'feature_family_replace': True})
+
+    # this is awful, switch to JSON POST
+    for key, value in request.params.iteritems():
+        if key.startswith(u'query_data[sorting]'):
+            key = key.replace('query_data[sorting]', '').replace('[', '').replace(']', '')
+            options.setdefault('sorting', {})
+            options['sorting'][key] = value